major bug fix : int -> float  angle

+++ controller/Odometry.h
@@ -83,9 +83,9 @@ protected:
     int m_totalTicksR = 0;/* Totat des tics droite */
     int m_totalDistance = 0; /* Total distance parcouru (odometry test purpose) */
 
-    int m_dDistance = 0; /*  Distance en mm parcouru en dt */
-    int m_dAvgTheta = 0; /* moyenne angle(t-1) et t en rad en dt */
-    int m_dTheta = 0; /* angle rad en dt */
+    float m_dDistance = 0; /*  Distance en mm parcouru en dt */
+    float m_dAvgTheta = 0; /* moyenne angle(t-1) et t en rad en dt */
+    float m_dTheta = 0; /* angle rad en dt */
     int m_lastTime = 0; /* dt */
